Use binary vertex shader generated by the libMali.so blob
Actually switching to an open source shader compiler would be a
better idea after it gets vertex shaders support in the long run.

Still the Mali Offline Shader Compiler v4.1.0 User Guide explicitly
mentions that the pre-compiled binary shaders are intended to be
distributed with applications. Also it is royalty free (with some
clauses against reverse engineering). As a matter of fact, no
reverse engineering is being done when developing this simple memory
testing tool. The bundled 3D demo is only used to additionally stress
the memory controller in order to expose stability issues.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>