A bit nicer handling of 'modprobe mali'
Suppress modprobe output and show our own error message in the
case of error. Also modprobe is now done before starting the
lima thread and we wait a bit longer before starting memtester.
So that it is now less likely to have console messages from
lima and memtester mixed together.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>