memtester: Try to guess if it was a read or write operation failing
Basically we just compare the buffers twice. And if the result is
reproducible, then it is likely that a corrupted value is already
in DRAM as a result of writing earlier. And if the result of
comparing buffers in not reproducible, then it was a read error
that happened during the compare operation itself.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>