A new lima-memspeed benchmark program
This program offers a number of different workloads, which try to
utilize as much memory bandwidth as possible when run in several
threads simultaneously. By combining the use of both CPU and GPU
at the same time, the memory bandwidth is maximized. Then the
measured memory bandwidth is reported to the console with a 5
seconds period.

The CPU based workloads are borrowed from the tinymembench program.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>