lima-memtester: Actively prevent screen from blanking
The commit 6d7fac197dfa4caaa45ea5fc7009dc40e8fde187 removed
screen blanking prevention from the lima driver. This was needed
for the lima-memspeed program (sometimes we want to keep the
screen blanked for memory bandwidth measurements).

Now we introduce screen blanking prevention in lima-memtester
back, because it is needed for the visual feedback (glowing red
background when the errors are found). The visual feedback is
needed for the users, who can't see the log messages on the
console.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>