lima-memtester: More accurately distinguish READ and WRITE errors
Read errors may be detected at exactly the same offsets on a
second try and because of this they may be incorrectly reported
as write errors. This patch just makes the error type check run
many times in a loop, reducing the probability of making a mistake.
And also crc32 is calculated for the whole comparison result in
addition to checking offsets.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>