Add Shapiro-Wilk test of normality
This might be not do much good for us. Normality tests need
a lot of samples to produce useful results. Also the Shapiro-Wilk
test does not like ties very much, while we are dealing with the
binned data. This patch also adds a hack to get rid of the ties.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>