genchart: Do exact multinomial test via R
https://en.wikipedia.org/wiki/Multinomial_test
https://cran.r-project.org/web/packages/XNomial/vignettes/XNomial.html

Basically, we want to verify whether the theoretical probabilities
which we assign to the DRAM clock frequencies (typically binned
in 24 MHz intervals) assuming the Gaussian distribution agree with
the experimental data.

If more than one set of clock frequencies is provided as input,
then the first one is used for training (calculating the theoretical
probabilities) and the second one is used for validation (checking
whether the observed clock frequencies from the second set fit
the theoretical model). If only one set is provided, then it is
validated against itself.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>