genchart: Remove the Shapiro-Wilk normality test and gnuplot support
The Shapiro-Wilk test is not very useful because having the binned
data is already problematic (we need to randomize it inside of each
interval as workaround, which is not pretty). The exact multinomial
test is a much better goodness of fit check in this case.

The gnuplot was used to show a cumulative distribution function,
which is not very interesting. Also uploading pictures to the wiki
is a little bit combersome. So remove the gnuplot support too.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>